client: Move client configuration model in separate file
