+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './client';
 export * from './lock';
+export * from './models';
 export * from './mongo-read-stream';
 export * from './mongo-utils';
 export * from './aggregation-utils';
