+++ src/models/index.ts
@@ -1,8 +1,11 @@
 export * from './aggregate.models';
 export * from './base-mongo-object.models';
 export * from './base-mongo-object-infos.models';
+export * from './class-type.models';
 export * from './historic.models';
 export * from './lock-field.models';
 export * from './maps.models';
+export * from './mongo-client-configuration.models';
+export * from './tag-options.models';
 export * from './update-many-at-once-options.models';
 export * from './update-many-query.models';
